update and delete user verification

+++ controllers/usersController.js
@@ -33,8 +33,17 @@ const getOneUser = asyncHandler(async (req, res, next) => {
   });
 });
 
+// user who logged in is the same user who get updated or deleted, or he is an admin
 const updateUser = asyncHandler(async (req, res, next) => {
   const id = req.params.id;
+
+  const userId = req.user.id;
+  const role = req.user.role;
+
+  if (userId !== id && role !== "admin") {
+    throw new AppError("you are not authorized", 403);
+  }
+
   const data = req.body;
 
   const updatedUser = await User.findByIdAndUpdate(id, data, { new: true });
@@ -49,6 +58,13 @@ const updateUser = asyncHandler(async (req, res, next) => {
 
 const deleteUser = asyncHandler(async (req, res, next) => {
   const id = req.params.id;
+  const userId = req.user.id;
+  const role = req.user.role;
+
+  if (userId !== id && role !== "admin") {
+    throw new AppError("you are not authorized", 403);
+  }
+
   const deletedUser = await User.findByIdAndDelete(id);
   if (!deletedUser) {
     throw new AppError("user not found", 404);
